# docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is the backend service for UBC UAS",
        "title": "GCOM Backend",
        "contact": {
            "name": "UBC UAS",
            "url": "https://ubcuas.com/",
            "email": "info@ubcuas.com"
        },
        "version": "1.0"
    },
    "host": "localhost:1323",
    "paths": {
...
        "/drone/home": {
            "post": {
                "description": "Updates the home waypoint",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Drone"
                ],
                "summary": "Updates the home waypoint",
                "parameters": [
                    {
                        "description": "Home Waypoint",
                        "name": "waypoints",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Waypoint"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/drone/land": {
            "get": {
                "description": "Tells Drone to land",
                "tags": [
                    "Drone"
                ],
                "summary": "Take off Drone",
                "responses": {
                    "200": {
                        "description": "Command issued successfully",
                        "schema": {
                            "type": "body"
                        }
                    },
                    "500": {
                        "description": "Command failed to be issued",
                        "schema": {
                            "type": "body"
                        }
                    }
                }
            }
        },
        "/drone/lock": {
            "get": {
                "description": "Stops drone movement while preserving existing queue",
                "tags": [
                    "Drone"
                ],
                "summary": "Halts drone in place while preserving queue",
                "responses": {
                    "200": {
                        "description": "Drone locked successfully",
                        "schema": {
                            "type": "body"
                        }
                    },
                    "500": {
                        "description": "Drone unable to lock (already locked?)",
                        "schema": {
                            "type": "body"
                        }
                    }
                }
            }
        },
        "/drone/queue": {
            "get": {
                "description": "Returns queue in Mission Planner",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Drone"
                ],
                "summary": "Returns queue in Mission Planner",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Waypoint"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Sends a queue in Mission Planner",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Drone"
                ],
                "summary": "Sends a queue in Mission Planner",
                "parameters": [
                    {
                        "description": "Array of Waypoint Data",
                        "name": "waypoints",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Waypoint"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/drone/rtl": {
            "post": {
                "description": "Tells Drone to return home and land",
                "tags": [
                    "Drone"
                ],
                "summary": "Returns to Home and Lands",
                "responses": {
                    "200": {
                        "description": "RTL command issued successfully",
                        "schema": {
                            "type": "body"
                        }
                    },
                    "500": {
                        "description": "RTL command encountered an error",
                        "schema": {
                            "type": "body"
                        }
                    }
                }
            }
        },
        "/drone/takeoff": {
            "post": {
                "description": "Tells Drone to takeoff",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Drone"
                ],
                "summary": "Take off Drone",
                "parameters": [
                    {
                        "description": "Takeoff Altitude",
                        "name": "altitude",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/drone/unlock": {
            "get": {
                "description": "Stops drone movement while preserving existing queue",
                "tags": [
                    "Drone"
                ],
                "summary": "Halts drone in place while preserving queue",
                "responses": {
                    "200": {
                        "description": "Drone unlocked successfully",
                        "schema": {
                            "type": "body"
                        }
                    },
                    "500": {
                        "description": "Drone unable to unlock (already unlocked?)",
                        "schema": {
                            "type": "body"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.Designation": {
            "description": "Describes a special purpose for a Waypoint",
            "type": "string",
            "enum": [
                "launch",
                "land",
                "obstacle",
                "payload"
            ],
            "x-enum-varnames": [
                "Launch",
                "Land",
                "Obstacle",
                "PayloadDesignation"
            ]
        },
...
        "models.Waypoint": {
            "description": "describes a location in GCOM",
            "type": "object",
            "required": [
                "alt",
                "id",
                "lat",
                "long",
                "name"
            ],
            "properties": {
                "id": {
                    "description": "To create a waypoint, ID of \"-1\" must be passed",
                    "type": "string",
                    "x-order": "1",
                    "example": "1"
                },
                "name": {
                    "type": "string",
                    "x-order": "2",
                    "example": "Alpha"
                },
                "lat": {
                    "type": "number",
                    "x-order": "3",
                    "example": 49.267941
                },
                "long": {
                    "type": "number",
                    "x-order": "4",
                    "example": -123.24736
                },
                "alt": {
                    "type": "number",
                    "x-order": "5",
                    "example": 100
                },
                "radius": {
                    "description": "Radius around waypoint where it is considered flown over",
                    "type": "number",
                    "x-order": "6",
                    "example": 10
                },
                "designation": {
                    "description": "Designation of waypoint, none by default",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Designation"
                        }
                    ],
                    "x-order": "7",
                    "example": "land"
                },
                "remarks": {
                    "type": "string",
                    "x-order": "8",
                    "example": "Task 1 Landing Zone"
                }
            }
        },
...
    }
}